basicas-colecciones/src/edu/unlam/progava/colecciones/Persona.java
package edu.unlam.progava.colecciones;

public class Persona implements Comparable<Persona> {

	private String nombre;
	private String ciudad;
	private int edad;
	private int sueldo;

	public Persona(String identificacion, int anios) {
		nombre = identificacion;
		edad = anios;
	}

	public String getNombre() {
		return nombre;
	}

	public void setNombre(String nombre) {
		this.nombre = nombre;
	}

	public int getEdad() {
		return edad;
	}

	public void setEdad(int edad) {
		this.edad = edad;
	}

	@Override
	public int compareTo(Persona otraPersona) {
		if (this.edad > otraPersona.getEdad())
			return 1;
		else if (this.edad < otraPersona.getEdad())
			return -1;
		return 0;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((ciudad == null) ? 0 : ciudad.hashCode());
		result = prime * result + edad;
		result = prime * result + ((nombre == null) ? 0 : nombre.hashCode());
		result = prime * result + sueldo;
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		Persona other = (Persona) obj;
		if (ciudad == null) {
			if (other.ciudad != null)
				return false;
		} else if (!ciudad.equals(other.ciudad))
			return false;
		if (edad != other.edad)
			return false;
		if (nombre == null) {
			if (other.nombre != null)
				return false;
		} else if (!nombre.equals(other.nombre))
			return false;
		if (sueldo != other.sueldo)
			return false;
		return true;
	}
	

}
